dom/imptests/html/dom/nodes/test_Document-getElementById.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Document.getElementById with an empty string</title>
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-document-getelementbyid">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<div id=""></div>
<script>
test(function() {
  assert_equals(document.getElementById(""), null)
  // XXX needs more tests
})
</script>
